+++ lib/index.ts
@@ -4,6 +4,10 @@ export * from './writer';
 export * from './errors';
 export * from './utils';
 
-export * from './sql/operators';
-export * from './sql/functions';
 export * from './sql/raw';
+export * from './sql/operators';
+export * from './sql/functions/type';
+export * from './sql/functions/time';
+
+// TODO Remove this once timeNow() is removed
+export * from './sql/functions';